# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Ecommerce Web Application built using Go Lang, PSQL, REST API following
    Clean Architecture.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Ecommerce Web API
  version: "1.0"
paths:
...
  /wishlist/:
    delete:
      consumes:
      - application/json
      description: User can remove all product items from wishlist
      operationId: empty-wishlist
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can remove all product items from wishlist
      tags:
      - Wishlist
    get:
      consumes:
      - application/json
      description: User view product items in wishlist
      operationId: view-wishlist
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can view items in wishlist
      tags:
      - Wishlist
  /wishlist/{id}:
    delete:
      consumes:
      - application/json
      description: User can remove product item from wishlist
      operationId: remove-from-wishlist
      parameters:
      - description: ID of the product item to be removed from wishlist
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can remove product item from wishlist
      tags:
      - Wishlist
    post:
      consumes:
      - application/json
      description: User can add product item to wishlist
      operationId: add-to-wishlist
      parameters:
      - description: ID of the product item to be added to wishlist
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can add product item to wishlist
      tags:
      - Wishlist
swagger: "2.0"